addEventsToWebsite() 이동 commonFunction.js -> init.js
